+++ src/timemanager/gui/controller/ProjectController.java
@@ -28,6 +28,7 @@ import javafx.scene.control.TableView;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.input.MouseButton;
 import javafx.scene.input.MouseEvent;
+import javafx.scene.layout.HBox;
 import javafx.stage.Stage;
 import javafx.util.Duration;
 import timemanager.TimeManagerStart;
@@ -82,6 +83,14 @@ public class ProjectController implements Initializable {
     private JFXComboBox<Client> selectClient;
     @FXML
     private JFXCheckBox billable;
+    @FXML
+    private HBox timeLoggerLink;
+    @FXML
+    private HBox clientsLink;
+    @FXML
+    private HBox usersLink;
+    @FXML
+    private HBox statisticsLink;
 
     /**
      * Initializes the controller class.
@@ -108,25 +117,50 @@ public class ProjectController implements Initializable {
         setupTimeline();
         
         initStartTimeline();
-    }    
+        
+        isAdmin();
+    }  
+    
+    private void isAdmin() {
+        if(lm.getLoggedInUser().getIsAdmin() == true) {
+            timeLoggerLink.setDisable(false);
+            timeLoggerLink.setVisible(true);
+            clientsLink.setDisable(false);
+            clientsLink.setVisible(true);
+            usersLink.setDisable(false);
+            usersLink.setVisible(true);
+            statisticsLink.setDisable(false);
+            statisticsLink.setVisible(true);
+            
+        } else {
+            usersLink.setDisable(true);
+            usersLink.setVisible(false);
+            statisticsLink.setDisable(true);
+            statisticsLink.setVisible(false);
+        }
+    }
 
     @FXML
     private void openTimeLogger(MouseEvent event) throws Exception {
+        timeline.stop();
         tms.set((Stage) (selectProject.getScene().getWindow()), "TimeLogger");
     }
     
     @FXML
     private void openClients(MouseEvent event) throws Exception {
+        timeline.stop();
         tms.set((Stage) (selectProject.getScene().getWindow()), "Client");
     }
 
     @FXML
     private void openUsers(MouseEvent event) throws Exception {
+        timeline.stop();
         tms.set((Stage) (selectProject.getScene().getWindow()), "User");
     }
 
     @FXML
     private void openStatistics(MouseEvent event) throws Exception {
+        timeline.stop();
         tms.set((Stage) (selectProject.getScene().getWindow()), "Statistics");
     }
 
@@ -319,7 +353,10 @@ public class ProjectController implements Initializable {
                 }
             }
         });
-        tcm.getItems().add(deleteItem);
+        
+        if(lm.getLoggedInUser().getIsAdmin() == true) {
+            tcm.getItems().add(deleteItem);
+        }
         
         projectTable.addEventHandler(MouseEvent.MOUSE_CLICKED, new EventHandler<MouseEvent>() {
