+++ src/timemanager/gui/controller/TaskController.java
@@ -9,8 +9,6 @@ import com.jfoenix.controls.JFXButton;
 import com.jfoenix.controls.JFXCheckBox;
 import com.jfoenix.controls.JFXComboBox;
 import java.net.URL;
-import java.time.LocalTime;
-import java.time.format.DateTimeFormatter;
 import java.util.Optional;
 import java.util.ResourceBundle;
 import java.util.logging.Level;
@@ -31,6 +29,7 @@ import javafx.scene.control.TableView;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.input.MouseButton;
 import javafx.scene.input.MouseEvent;
+import javafx.scene.layout.HBox;
 import javafx.stage.Stage;
 import javafx.util.Duration;
 import timemanager.TimeManagerStart;
@@ -81,6 +80,14 @@ public class TaskController implements Initializable {
     private TableColumn<Task, String> description;
     @FXML
     private TableView<Task> taskTable;
+    @FXML
+    private HBox timeLoggerLink;
+    @FXML
+    private HBox clientsLink;
+    @FXML
+    private HBox usersLink;
+    @FXML
+    private HBox statisticsLink;
 
     /**
      * Initializes the controller class.
@@ -105,7 +112,28 @@ public class TaskController implements Initializable {
         setupTimeline();
         
         initStartTimeline();
+        
+        isAdmin();
     }  
+    
+    private void isAdmin() {
+        if(lm.getLoggedInUser().getIsAdmin() == true) {
+            timeLoggerLink.setDisable(false);
+            timeLoggerLink.setVisible(true);
+            clientsLink.setDisable(false);
+            clientsLink.setVisible(true);
+            usersLink.setDisable(false);
+            usersLink.setVisible(true);
+            statisticsLink.setDisable(false);
+            statisticsLink.setVisible(true);
+            
+        } else {
+            usersLink.setDisable(true);
+            usersLink.setVisible(false);
+            statisticsLink.setDisable(true);
+            statisticsLink.setVisible(false);
+        }
+    }
     
     public void setProject(Project project) {
         this.project = project;
@@ -114,21 +142,25 @@ public class TaskController implements Initializable {
 
     @FXML
     private void openTimeLogger(MouseEvent event) throws Exception {
+        timeline.stop();
         tms.set((Stage) (selectProject.getScene().getWindow()), "TimeLogger");
     }
     
     @FXML
     private void openClients(MouseEvent event) throws Exception {
+        timeline.stop();
         tms.set((Stage) (selectProject.getScene().getWindow()), "Client");
     }
 
     @FXML
     private void openUsers(MouseEvent event) throws Exception {
+        timeline.stop();
         tms.set((Stage) (selectProject.getScene().getWindow()), "User");
     }
 
     @FXML
     private void openStatistics(MouseEvent event) throws Exception {
+        timeline.stop();
         tms.set((Stage) (selectProject.getScene().getWindow()), "Statistics");
     }
 
@@ -291,7 +323,10 @@ public class TaskController implements Initializable {
                 }
             }
         });
-        tcm.getItems().add(deleteItem);
+        
+        if(lm.getLoggedInUser().getIsAdmin() == true){
+            tcm.getItems().add(deleteItem);
+        }
         
         taskTable.addEventHandler(MouseEvent.MOUSE_CLICKED, new EventHandler<MouseEvent>() {
